Drop redundant (double) cast in GalleryModel.CalculateUsage
cs/useless-cast-to-self introduced by my prior edit. The
multiplicand 1000000.0 is already double, so TotalGallerySize /
(maxSizeMB * 1000000.0) is double and the outer cast is redundant.

+++ Memtly.Core/Memtly.Core/Models/Database/GalleryModel.cs
@@ -17,7 +17,7 @@ namespace Memtly.Core.Models.Database
 
         public string CalculateUsage(long maxSizeMB = long.MaxValue)
         {
-            return ((double)(TotalGallerySize / (maxSizeMB * 1000000.0))).ToString("0.00%");
+            return (TotalGallerySize / (maxSizeMB * 1000000.0)).ToString("0.00%");
         }
     }
 }
